+++ CHANGELOG.md
@@ -67,6 +67,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix issue that domain globs not supported during wwctl node delete. #1449
 - Fix overlay permissions in /root/ and /root/.ssh/. #1452
 - Return non-zero exit code on container sub-commands #1437
+- Return non-zero exit code on profile sub-commands #1435
 
 ## v4.5.8, 2024-10-01